Finished data filtering script rewrite

+++ filtering/bulbapedia_filter.py
@@ -13,12 +13,16 @@ def filter_pokemon(pokemon: Pokemon):
     path = f"data/bulbapedia/{get_pokemon_directory(pokemon)}"
     directory = os.path.join(dir, path)
 
+    exclusion_terms = ["funko", "komaplaza", "art_life", "binder", "box", "volume", "card", "coin", "print", "block", "paper", "quest", "trophy", "shirt", "sleeve", "deck"]
+
     # Count images in directory
     image_files = [file for file in os.listdir(directory) if file.lower().endswith((".png", ".jpg", ".jpeg"))]
     image_count = len(image_files)
 
+    dry_run = False
+
     # Define thresholds
-    minimum_width = 94
+    minimum_width = 80
     minimum_height = 80
     transparency_threshold = 0.8
     transparency_minimum = 0.00
@@ -33,46 +37,59 @@ def filter_pokemon(pokemon: Pokemon):
 
             # Check for small images (less than 100px)
             if width < minimum_width or height < minimum_height:
-                os.remove(image_path)
+                if not dry_run:
+                    os.remove(image_path)
                 deleted_count += 1
+                print("Deleting for size")
                 continue
 
             # Remove palette or greyscale images
-            if img.mode == "P" or img.mode == "L":
-                os.remove(image_path)
+            if img.mode == "P" or img.mode == "L" or img.mode == "LA":
+                if not dry_run:
+                    os.remove(image_path)
                 deleted_count += 1
+                print("Deleting for format")
                 continue
-
-            # Remove non-transparent images
-            """if img.mode == "RGB":
-                os.remove(image_path)
-                deleted_count += 1
-                continue
-            """
 
             # Remove images with titles that suggest the Pokemon is shiny
             filename = image_file.lower().split(".")[-2]
             if filename.endswith("_s") or "shiny" in filename:
-                os.remove(image_path)
+                if not dry_run:
+                    os.remove(image_path)
                 deleted_count += 1
                 continue
 
             # Check for transparent pixels if the image has an alpha channel
-            if img.mode == "RGBA":
-                transparent_pixels = 0
-                total_pixels = width * height
-                pixels = img.getdata()
+            transparent_pixels = 0
+            total_pixels = img.width * img.height
+            pixels = img.getdata()
 
+            if img.mode == "RGBA":
                 for pixel in pixels:
-                    if pixel[3] == 0:  # Alpha channel at 0 means fully transparent
+                    if pixel[3] == 0:
                         transparent_pixels += 1
 
+                # Calculate ratios
                 transparency_ratio = transparent_pixels / total_pixels
-                if transparency_ratio > transparency_threshold or transparency_ratio < transparency_minimum:
-                    os.remove(image_path)
-                    deleted_count += 1
 
-    print(f"{pokemon.name} filtered: started at {image_count} ended at {image_count - deleted_count} (deleted {deleted_count})")
+                # Check thresholds
+                if transparency_ratio > transparency_threshold or transparency_ratio < transparency_minimum:
+                    if not dry_run:
+                        os.remove(image_path)
+                    deleted_count += 1
+                    continue
+        
+        # Check image name to remove unwanted names
+        for exclusion in exclusion_terms:
+            if exclusion in image_file.lower() and exclusion not in pokemon.name.lower():
+                if not dry_run:
+                    os.remove(image_path)
+                deleted_count += 1
+                print(f"Deleting for exclusion term {exclusion} for {image_file}")
+                break
+
+    if deleted_count > 0:
+        print(f"{'Dry Run: ' if dry_run else ''} {pokemon.name} filtered: started at {image_count} ended at {image_count - deleted_count} (deleted {deleted_count})")
 
 
 def filter_form(pokemon: Pokemon, base_form: Pokemon):
@@ -119,19 +136,25 @@ def main() -> None:
     end_at = 1030
     pokemon_data: list[Pokemon] = get_pokedex(False, start_at, end_at)
 
+    # Set when sorting has already been done and should not be repeated
+    has_been_sorted_into_forms = True
+
     base_form: Pokemon | None = None
 
     for pokemon in pokemon_data:
         # print(get_pokemon_directory(pokemon))
-        if pokemon.standard:
+        if has_been_sorted_into_forms:
             filter_pokemon(pokemon)
-
-            if base_form:
-                base_form = None
-            if pokemon.alternate_count > 0:
-                base_form = pokemon
         else:
-            filter_form(pokemon, base_form)
+            if pokemon.standard:
+                filter_pokemon(pokemon)
+
+                if base_form:
+                    base_form = None
+                if pokemon.alternate_count > 0:
+                    base_form = pokemon
+            else:
+                filter_form(pokemon, base_form)
 
 
 if __name__ == "__main__":
